Add citadel generation, allow for non-grid locations
The citadel is represented by a GameObject attached to the GenerationManager: it will be scaled with NodeSize IF there is a GridManager present. Ensure the base prefab used is scaled properly with buildings at NodeSize 1. The generation now accounts for locations positioned outside of the grid, meaning that pre-placed locations can be set outside the grid and generate their contents (except for roads, as their generation is heavily grid-reliant)